* 0587-ThumbnailCreator-zach_mullen:
  BUG: refs #587. Must use upgrade script instead of 1.0.0 defs
  ENH: refs #587. Add large thumbnail view on an item from thumbnailcreator
  BUG: refs #0587. Removing echos as they break the web-api.
  BUG: refs #587. Fix case where thumbnail creation fails synchronously
  BUG: refs #587. Fix array reference in thumbnail creator

+++ modules/thumbnailcreator/controllers/components/ImagemagickComponent.php
@@ -32,7 +32,10 @@ class Thumbnailcreator_ImagemagickComponent extends AppComponent
     {
     $modelLoader = new MIDAS_ModelLoader;
     $itemModel = $modelLoader->loadModel('Item');
-    $item = $itemModel->load($item['item_id']);
+    if(is_array($item))
+      {
+      $item = $itemModel->load($item['item_id']);
+      }
     $revision = $itemModel->getLastRevision($item);
     $bitstreams = $revision->getBitstreams();
     if(count($bitstreams) < 1)
@@ -56,24 +59,23 @@ class Thumbnailcreator_ImagemagickComponent extends AppComponent
       }
     catch(phMagickException $exc)
       {
-      echo $exc->getMessage();
+      return;
       }
     catch(Exception $exc)
       {
-      echo $exc->getMessage();
+      return;
       }
 
     if(file_exists($pathThumbnail))
       {
       $oldThumbnail = $item->getThumbnail();
-      if(!empty($oldThumbnail))
+      if(!empty($oldThumbnail) && file_exists($pathThumbnail))
         {
         unlink($oldThumbnail);
         }
       $item->setThumbnail(substr($pathThumbnail, strlen(BASE_PATH) + 1));
       $itemModel->save($item);
       }
-    return;
     }
 
   /**
@@ -105,10 +107,12 @@ class Thumbnailcreator_ImagemagickComponent extends AppComponent
       {
       mkdir($tmpPath);
       }
-    $destination = $tmpPath.'/'.rand(1, 1000).'.jpeg';
+    $num = '0';
+    $destination = $tmpPath.'/'.$num.'.jpeg';
     while(file_exists($destination))
       {
-      $destination = $tmpPath.'/'.rand(1, 1000).'.jpeg';
+      $num++;
+      $destination = $tmpPath.'/'.$num.'.jpeg';
       }
     $pathThumbnail = $destination;
